requirements_1.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
drf-yasg==1.21.10
celery==5.5.3
pika==1.3.2
django-environ==0.12.0
